KA-56 - Rewrite actions and utils to typescript
