+++ onboarding-task/src/actions/IAction.ts
@@ -1,0 +1,6 @@
+interface IAction {
+  type: string;
+  payload: any;
+}
+
+export { IAction };
