--- onboarding-task/src/actions/addItemFactory.js
@@ -1,17 +1,0 @@
-import { ADD_ITEM } from './actionTypes.js';
-import { Item } from '../models/Item.ts';
-
-function addItem(text, generateNewId) {
-  return {
-    type: ADD_ITEM,
-    payload: {
-      item: new Item({ id: generateNewId(), text }),
-    },
-  };
-}
-
-function addItemFactory(idGenerator) {
-  return text => addItem(text, idGenerator);
-}
-
-export { addItemFactory };
