+++ onboarding-task/src/actions/editedItemsActionCreators.ts
@@ -1,6 +1,7 @@
-import { START_EDIT_ITEM, STOP_EDIT_ITEM } from './actionTypes.js';
+import { START_EDIT_ITEM, STOP_EDIT_ITEM } from './actionTypes';
+import { IAction } from './IAction';
 
-function startEditItem(id) {
+function startEditItem(id: string): IAction {
   return {
     type: START_EDIT_ITEM,
     payload: {
@@ -9,7 +10,7 @@ function startEditItem(id) {
   };
 }
 
-function stopEditItem(id) {
+function stopEditItem(id: string): IAction {
   return {
     type: STOP_EDIT_ITEM,
     payload: {
